Replace s3/local writer with songbird writer
This change adds the ability to send dataset profiles to songbird. It
adds a bunch of new env variables that have to be supplied for org and
the songbird api. I'll have to make this a bit more resilient before
using it though. It was pretty safe to depend on s3 being up to transmit
the dataset profiles in the old implementation but I think it would be
prudent to spend a bit of time working through failure scenarios and
debugging issues.