Fix error case that resulted in zombie process
Throwing at the top level before javalin started resulted in a zombie
process, kept alive by our schedule'd job for writing songbird files.
This ensures that any top level errors will shutdown the process so that
Docker can see the actual status of our main command correctly, which
allows it to manage our container with the --restart option.

Instead of using a process manager, I'm recommending we use docker to
manage the process with --restart=on-failure:3. My research online
indicates that using a process manager for a single process in a
container is an anti pattern and it obfuscates the status of the main
process from Docker's point of view (or other orchestrators). We should
focus on making sure that our app correctly terminates and background
threads play nicely so that the the exit code is correct and shutdown
happens at the right time. That's require regardless of what manages our
process.