Manage javalin with supervisord
This change adds supervisord to manage our main app process. It adds a
bit of bulk to the container (445.7MB up from ~250MB), but we do get
some advantages. Specifically,

- We can ensure the app process is restarted upon error without having
  to make sure that airspace starts the container with the right command
  line arguments.
- We can now restart the rest server without taking down the container,
  which is really nice for debugging scenarios that would have resulted
  in the container getting destroyed, which would have taken the file
  storage with it on heroku.

supervisord is currently set up to pretty much pass through all of the
logs and std io as-is so it will look just like it did without
supervisord, which is nice for getting a hold of logs if we need them.