Profiles stored in persistent storage, not in memory + lock removal
This change introduces a new abstraction (PersistentMap) that acts as
the storage layer for the dataset profiles. It's a map-like interface
that wraps Kotlin coroutines/actors to ultimately write values to sqlite
for storage after each write. This gives the container the ability
restart without losing all of the dataset profile information in the
in-memory map.

A side effect of this change is that we no longer need to handle
concurrency with locks and java concurrent classes. Everything plays
nice with Kotlin's concurrency primitives. The only thing that I left is
the @Volatile timestamp since it was a pretty small detail and its only
set from a single thread.