Decouple uploading profiles from processing requests
The initial version of the container used a single data structure that
serialized all actions for both uploading profiles to whylabs and for
enqueueing request data to be uploaded. That meant that requests would
periodically pause for as long as it took to upload pending profiles to
whylabs.

This iteration adds a second queue structure that is used for enqueueing
requests independent of the structure that stores pending profiles for
upload. Enqueued requests are asynchronously moved into the upload
structure without blocking client requests.

Concurrency is handled without locks by serializing access to both the
queue (used for handling requests quickly) and the map (used for storing
pairs for tag-profile waiting for upload), so high concurrency does
incur some overhead. If we want to avoid this in the future then we can
add a perf flag that allows requests to finish without guaranteeing that
they were handled successfully first, since they're very likely to be
handled.