Add configuration option to send empty profiles when no data comes in
This new config option, EMPTY_PROFILE_DATASET_IDS, allows users to
specify a list of dataset ids that we should write profiles for every
period of time, regardless of whether or not we actually got any data
into the container. We don't have any way of doing this without knowing
what dataset ids they're using so we need them to tell us.