Set package versions and remove user specification
The package versions will ensure we don't pick up any new startup
behavior between builds. The removal of the user field brings it in line
with how we used to do things before migrating to github CI. I had added
this as a side effect of an error I was getting that I don't seem to be
getting anymore so if we don't need it then I'll just remove it.