Add basic Kafka consumer functionality
This change introduces a new mode of operation for the container: kafka
consumer mode. This mode can be utilized instead of, or, in addition to
the rest container mode that already exists. In this mode, the container
will act as a kafka consumer on a configured cluster's topic, folding
all of the data in that topic into whylogs profiles the same way that it
handles rest requests.

There are a few differences between kafka and rest mode:
- Kafka mode doesn't utilize any container based buffering/queueing
systems while the rest container mode does buffer requests. It just
relies on kafka for that.
- Kafka mode assumes (for now) a certain format for the messages: Maps
of key value pairs that can be optionally nested. Each message is
assumed to be a single datapoint. REST mode allows you to specify
multiple data points per request.

This change also updated the configuration format to add namespaces for
the different logical groups of config, now that there are some that
apply only to kafka, and some that apply only to REST.