Fix issue that caused errors to stop processing requests
The queue abstraction that we use ontop of sqlite wasn't able to recover
from errors when sqlite threw. Now it recovers.

We were getting a SQLite BUSY error which is technically possible since
we're allowing concurrent writes in the queue via pop/push, just rare
that the pop/push would happen at the same time. In anycase, we retry
for errors now and when things don't work out we won't die forever.

I refactored the layout of the queue actors so that they run on
their own threads when they're running in concurrent mode as well so
there is only one connection per thread at most.

Another useful change here is to the log format. I enabled coroutine
debugging and updated the format so that we'll see which coroutine
context, coroutine id, and thread id each message came from. That will
be very  useful when debugging some supposedly impossible error that
we're bound to get.

This also adds retries to the api for the queue/map as the mechanism
that compensates for the potential of errors in the write layer. This
was nice for actually testing the retry logic as well.

Changed the default increment size for draining queued requests to be
100, down from All. All may perform better in some situations but the
risk of OOMing the box is too big.