Add basic Kafka consumer functionality
This change introduces a new mode of operation for the container: kafka
consumer mode. This mode can be utilized instead of, or, in addition to
the rest container mode that already exists. In this mode, the container
will act as a kafka consumer on a configured cluster's topic, folding
all of the data in that topic into whylogs profiles the same way that it
handles rest requests.

There are a few differences between kafka and rest mode:
- Kafka mode doesn't utilize any container based buffering/queueing
systems while the rest container mode does buffer requests. It just
relies on kafka for that.
- Kafka mode assumes (for now) a certain format for the messages: Maps
of key value pairs that can be optionally nested. Each message is
assumed to be a single datapoint. REST mode allows you to specify
multiple data points per request.

This change also updated the configuration format to add namespaces for
the different logical groups of config, now that there are some that
apply only to kafka, and some that apply only to REST.

This is also a sizable perf increase to the container as a whole. There
were several changes to sqlite: reusing a single connection for a long
time, reusing prepared statements, etc. These net out to a large
increase in p50 latency and tps. Exact figures depend on hardware of
course, but throughput increased by 5x-6x, which made the request
buffering optimization that I put in totally unnecessary, which is why
it is now disabled by default. That has the added benefit of simplifying
the architecture and removing the piece that potentially required
unbound hdd space because I hadn't built any mechanism to limit request
queueing.