Add lightweight integration tests
These aren't full end to end tests by any means but they are capable of
running blackbox tests on the runtime system. For now, they use a
special dedicated file writer as the container output, rather than s3 or
whylabs, in order to make assertions against the generated profiles
based on the container requests. These are going to be run as part of
the CI before pushes as well so we'll be able to catch issues that break
the runtime here as we add more tests.