Checking in prototype and findings
Was trying to determine the best way to enable stack trace recovery and
apm in the container. I was experimenting with sentry.io and datadog but
there were a couple of things that I didn't like:

- Pricing for both of them is a little steep for what we want. Datadog
has per host charging and this is a free (beer) open source container.
- We would have to find some way of getting the api key into the
container for our account that doesn't require checking it into source.
- We would probably vend it out in exchange for an API key from whylabs
so we could control access, or just proxy it through whylabs and hide
that detail entirely.
- Sentry charges by transaction. They only include the transactions
(requests) you send to them in their percentile calculations, which is
very lame. We would have to enable sampling and compromise our apm
because of their pricing when they should just be profiling on hosts and
aggregating those profiles after.

Some other interesting notes:
- Sentry's client is open source and can be configured to send events
anywhere