Fix issue uploading many profiles at once + scheduled job run
Two main issues here.

First, we use ExecutorService.scheduleWithFixedDelay to schedule
periodic uploads to whylabs. If that upload throws then the
ExecutorService will consider the schedule cancelled and stop uploading
profiles. The fix for this to swallow errors around uploads since
they're just rethrows at that point, they're already handled by keeping
the profile in storage.

Second, there is a hard coded 10 second timeout in the api that handles
processing profile storage. If there were enough profiles stored to make
the total upload time > 10 seconds then the processing attempt would
cancell the upload assuming the entire thing timed out and keep all of
the profiles in storage for later. Whatever got uploaded would remain in
storage and that would happen forever.

The fix for the second one was to expose a second api that allows
profiles to be handled one at a time, instead of as a giant collection.
That means that the 10 second timeout can apply to a single profile
upload, rather than all of them, and the things that were uploaded could
be safely removed from storage if something were to go wrong so we would
eventually handle everything.