Simplify kafka consumer shutdown behavior and fix potential bug
The way I had it before, while cooler than it was, technically wasn't
correct. By shutting down the consumer in the poll loop I risked the
`close` being called on a separate thread, since each iteration called
`yeild` and potentially swapped threads to another in Dispatchers.IO.
Now, each consumer has an actual dedicated thread that it runs on with a
dedicated context and closing happens in the shutdown hook's Thread by
switching over to that context in `runBlocking` and executing the close
there.

The KotlinConsumer doesn't need to be a CoroutineScope at all anymore
and the worker launches don't need to be launched themselves. The
consumer is totally self contained in its scope with its shutdown logic
and the shutdown hook waits for the close.