Allow specifying upload cadence and default to IN_MEMORY
The sqlite implementation of the profile store has this bad perf issue
that happens when the container runs long enough so I'm going to default
the store to in-memory instead. To compensate for the potential to lose
data via javalin crashes (already a very rare event), I added an option
to separately specify the cadence to upload profiles.

Now, you can have an hourly model whose profiles you upload every 5, 10,
30 minutes, for example. That mitigates the max amount of data you
might lose for a transient crash and doesn't take the massive perf hit
that you currently take defaulting to sqlite.