Several updates to improve debugging, small fixes
First, the small fixes:
- Make IGNORED_KEYS work for the rest interface's `multiple` field as
well. It worked only for Kafka and the `single` field before.
- Trim API key whitespace. I ran into an issue where an API key wasn't
working because of this, which seemed silly.

The biggest update for debugging was the addition of a separate actor
that tracks a bunch of information that I find useful for debugging
issues, especially remotely. This debug state can be retrieved from logs
after the new API `logDebugInfo` is called, or after a shutdown. See the
`DebugInfo`file/class for the specific info collected.